requirements.txt
# Web Interface
streamlit==1.53.0

# AI Core & Orchestration
langchain==0.2.6
langchain-openai==0.1.14
langchain-community==0.2.6
langchain-chroma==0.1.4
langchain-text-splitters==0.2.0

# Data Processing & Storage
chromadb==0.5.3
pydantic==2.12.5
pydantic-settings==2.10.1
pypdf==5.1.0
python-dotenv==1.1.1

# Tools
duckduckgo-search==8.1.1
requests==2.32.5
beautifulsoup4==4.12.3

# Web Scraping
selenium==4.40.0
undetected-chromedriver==3.5.5

# Development
python-dotenv==1.1.1